virtio: rename virtq to virtq_device
